Minor traction control fixes

+++ Dashboard_Firmware_Attempt2/Core/Inc/traction_control.h
@@ -12,15 +12,15 @@
 #ifndef SRC_TRACTION_CONTROL_H_
 #define SRC_TRACTION_CONTROL_H_
 
-extern volatile uint16_t TC_control_var;
+extern volatile float TC_control_var;
 extern volatile uint16_t TC_torque_adjustment;
 extern volatile float current_slip_ratio;
 
-extern volatile uint16_t pid_error;
-extern volatile uint16_t prev_pid_error;
+extern volatile float pid_error;
+extern volatile float prev_pid_error;
 
-extern volatile uint16_t integral;
-extern volatile uint16_t derivative;
+extern volatile float integral;
+extern volatile float derivative;
 
 void traction_control_PID(uint32_t fr_wheel_speed, uint32_t fl_wheel_speed);
